mlist converted to module
Purpose is making running certain alloc operations which can't easily be
done by static initialization. This is starting to resemble OOP Turns out
putting init functions in .init in section is not so easy to do
transparently as .init need to contain actual code and not a list of
function pointers (as in the case with .ctors/.dtors). I.e. there should be
function pointers there, but each one should be preceded by a call, which
becomes architecture specific.

One could had used __attribute__ ((__constructor__)) instead, but chose
using linker _init/_fini instead. Unfortunately this only works for shared
libraries so might reconsider later (as .so's makes this small project more
cumbersome to handle).

(Tinker has another much more elegant method which resembles more to the
.ctors/.dtors. Linux kernel just puts __init prefixed function in .init
section, not sure why that works as it shouldn't)