libmlist: Structure for list-container finally in place
Both data and code is starting to flux less. Good enough for continue
without finalizing it as libmlist is an ambitious project on its own with
support for reentrant calling and multi-threaded access Can't be fully as
references (i.e. pointers) can't be safe without working with copies and I
don't want that as list access needs to be fast lighting fast. Support will
work on protecting internal structure during creation and deletion, which is
perfectly good as long as one understands that additional synchronisation is
needed depending on use-case.

Embryonic functionality of the sampler engine, i.e. threads created by
create_executor.

Category: development