Memory corruption fixed
Beginners error when hacking too fast. This correction proved hard as the
tools for multi-threaded code are really sucky. Error was pure logic however
and Tinker in simulated mode would had  played a role here as it has the
advantage of that all threads stop when one hit a break-point. As it becomes
a user-mode scheduling library a.k.a. Sun-sparc theading library, all
threads execute in the shope of one process from the host-OS perspective.
Having that ability is otherwise available to HW debuggers only.

There are two errors actually. The one solved and the original problem,
which is that the lockstep doesn't work. Master thread doesn't block where
it should (end barrier) until all workers have completed. This happens only
when code executes full throttle. Therefore visual aids were introduced in
this commit in form of abilities (changed output format) for
driveGnuPlotStreams as it turns out that feedgnuplot gets confused about
streams missing and can start plotting on the wrong one (!). This
becomes a weakness of aosp.tzplots.sh too and might seriously mislead analyse.

Category: bugfix, analyze, development