Shamelessly adding files from Tinker
I can do that because I'm the Copyright-holder of both :-)

These files will be transformed into debug-wrappers as debugging RT-code
with weak tools is a night-mare. This is an alternative to pull in the
complete Tinker.

The files  will add the control needed to know what the kernel "should" be
doing, thereby make it much easier to identify states without digging into
kernel only because you need to debug in user-land.

To use: Nothing

All will be handled by Makefiles, who replace selected pieces of the system
system with the wrappers.  Wrappers in turn do what's needed, and cascade
calls further to the original functions. Easy-peasy...

Category: administration