Adding external libraries the right way
By testing for any function known to possibly be part of a library, CMake
implicitly detects if library is needed or not. This is probably not the
use-case CHECK_LIBRARY_EXISTS was designed for but works well for targets
that has the required functions, but else-where.

Alone this will not suffice for targets that don't have the functions. For
example "capabilities" (-lcap) exists only for Linux and Android, even
though -lcap isn't needed as external library (i.e. library is either
"built-in" or provided else-where).

Todo: For BSD & derivatives additional handling needs to be done in code.

This should obsolete any hackish comments in the commit-log.